Move command strategy to individual file
