+++ guet/commands/get/get_factory.py
@@ -5,7 +5,8 @@ from guet.commands.command_factory import CommandFactoryMethod
 from guet.commands.get.committer_printing_strategy import CommitterPrintingStrategy
 from guet.commands.get.invalid_identifier_strategy import InvalidIdentifierStrategy
 from guet.commands.help_message_strategy import HelpMessageStrategy
-from guet.commands.strategy_command import StrategyCommand, CommandStrategy
+from guet.commands.strategy_command import StrategyCommand
+from guet.commands.strategy import CommandStrategy
 from guet.config.committer import Committer
 from guet.config.get_committers import get_committers
 from guet.config.get_current_committers import get_current_committers
