+++ guet/commands/get/invalid_identifier_strategy.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from guet.commands.strategy_command import CommandStrategy
+from guet.commands.strategy import CommandStrategy
 from guet.settings.settings import Settings
