+++ guet/commands/lambda_strategy.py
@@ -1,6 +1,6 @@
 from typing import List, Callable
 
-from guet.commands.strategy_command import CommandStrategy
+from guet.commands.strategy import CommandStrategy
 from guet.settings.settings import Settings
