+++ guet/commands/too_many_args.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from guet.settings.settings import Settings
-from guet.commands.strategy_command import CommandStrategy
+from guet.commands.strategy import CommandStrategy
 
 
 class TooManyArgsStrategy(CommandStrategy):
